Add files via upload
For this demonstration, you will use the bank marketing data set. So, let’s try and understand the problem statement to utilise the information available in the best possible way and proceed in the right direction as per the business problem at hand.

 

So, a bank ran a marketing campaign in the past and has obtained data pertaining to nearly 11,000 customers, which includes variables such as their age, jobs, bank balance, education, loan status and so on. Based on this data, the bank wants to develop its future strategies based on the insights that it drew from the previous campaign and improve for the next campaign so that more customers agree to open term deposits with the bank.

 

Hence, ‘deposit’ is the target variable here. A ‘Yes’ in the ‘deposit’ column indicates that the campaign was successful and the customer agreed to open a term deposit account with the bank. In contrast, a ‘No’ in the ‘deposit’ column indicates that the campaign was not very successful and the customer could not be convinced to open a term deposit account.

 

Essentially, the bank wants to:

Build a model that quantitatively relates to the success of the marketing campaign with variables such as job, marital status, education, bank balance, etc.
Identify the features of the data set that affect the successful conversion of customers.
To know the accuracy of the model, i.e., how well these variables predict the success of the campaign.